Return errors more often
At Zed, by far the most common panics we experience across the userbase
are from Linux users who crash during a `draw` call in the Blade
renderer.

We'd like to be able to detect that something has gone wrong and show a
notification via dbus to help people understand the problem and debug.