Optional fragment shader (#228)
Closes #226 

This changes the declaration of the fragment function from being
mandatory to instead be optional.

Vulkan and Metal have been tested in the sense that bunnymark with
fragment shader set to None compiles and runs. Upon inspection of vulkan
run of the bunnymark example without fragment specified, I observed that
the vertex shader is executed correctly but renderdoc states there is no
fragment shader stage as expected.