Adds support for external buffers and images (#252)
The use I'm using it for is to have external images using a Win32
handle. Because this is very platform specific and niche I instead made
a way to import self-made images and memory.

Specifically i exposed the vulkan device, the `Allocation` struct,
`allocate_memory` and `free_memory`.
I added 2 functions to the vulkan context for importing memory and
textures.
And added `select_memory_type` as a utility function.

I also added a way to add extensions when creating the context using a
new struct `PlatformSpecificContextDesc`.
Which is used by the new function `init_with_platform_desc`. The
original init function is still the same so it doesn't break anything.